again notification, with workmanager (no more alertManager) and service with countdown + data from firebase collection preference for when and what day send notifications

+++ app/src/main/java/com/uimainon/go4lunch/controllers/RecyclerView/ListRestaurantAdapter.java
@@ -88,7 +88,7 @@ public class ListRestaurantAdapter extends RecyclerView.Adapter<ListRestaurantAd
         Result result = mGooglePlaceData.get(position);
         holder.itemNameRestaurant.setText(result.getName());
         holder.itemAdressRestaurant.setText(result.getVicinity());
-        int nbrStars = searchNbrStrars(nbrWorker/result.getNbrVote());
+        int nbrStars = searchNbrStrars((double)nbrWorker/result.getNbrVote());
         showNumberStarOnView(holder, nbrStars);
         holder.itemAdressRestaurant.setText(result.getVicinity());
         int castDistance = (int) result.getDistance();
